refactor(client): remove dead HttpRequestClient and relocate one-off utility classes
- Delete org.apache.hudi.util.HttpRequestClient (zero usages repo-wide)
- Move HilbertCurveUtils to hudi-spark-client org.apache.hudi.sort, next to its
  sole consumer SpaceCurveSortingHelper; hilbert-curve dependency moves with it
  (dropped from flink/kafka-connect bundle shade lists, no longer in their tree)
- Move DistributedRegistryUtil to hudi-spark-client org.apache.hudi.metrics,
  next to DistributedRegistry (sole consumer is SparkRDDWriteClient)
- Move ExecutorFactory to org.apache.hudi.execution where all its consumers live
- Removes the single-class org.apache.hudi.optimize package

Behavior-preserving; no moved class is @PublicAPIClass, config-FQN-referenced,
or Kryo-registered.
