+++ hudi-client/hudi-spark-client/pom.xml
@@ -30,6 +30,13 @@
   <packaging>jar</packaging>
 
   <dependencies>
+    <!-- Hilbert Curve -->
+    <dependency>
+      <groupId>com.github.davidmoten</groupId>
+      <artifactId>hilbert-curve</artifactId>
+      <version>0.2.2</version>
+    </dependency>
+
     <!-- Scala -->
     <dependency>
       <groupId>org.scala-lang</groupId>
