+++ hudi-client/hudi-spark-client/src/main/java/org/apache/hudi/client/SparkRDDWriteClient.java
@@ -48,7 +48,7 @@ import org.apache.hudi.table.HoodieSparkTable;
 import org.apache.hudi.table.HoodieTable;
 import org.apache.hudi.table.action.HoodieWriteMetadata;
 import org.apache.hudi.table.upgrade.SparkUpgradeDowngradeHelper;
-import org.apache.hudi.util.DistributedRegistryUtil;
+import org.apache.hudi.metrics.DistributedRegistryUtil;
 
 import com.codahale.metrics.Timer;
 import lombok.AllArgsConstructor;
