+++ hudi-client/hudi-spark-client/src/main/java/org/apache/hudi/sort/HilbertCurveUtils.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.hudi.optimize;
+package org.apache.hudi.sort;
 
 import org.davidmoten.hilbert.HilbertCurve;
