+++ packaging/hudi-kafka-connect-bundle/pom.xml
@@ -107,7 +107,6 @@
 
                                     <include>com.lmax:disruptor</include>
                                     <include>com.github.davidmoten:guava-mini</include>
-                                    <include>com.github.davidmoten:hilbert-curve</include>
                                     <include>com.github.ben-manes.caffeine:caffeine</include>
                                     <include>com.yammer.metrics:metrics-core</include>
                                     <include>com.beust:jcommander</include>
